Restructure project to allow exapnding number of endpoints
